+++ src/test/java/ru/praktikumservices/qascooter/OrdersListTests.java
@@ -1,5 +1,6 @@
-package ru.praktikum_services.qa_scooter;
+package ru.praktikumservices.qascooter;
 
+import api.client.OrderClient;
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.Response;
